feat: Java 17 compile support
Libraries are compiled with Java 17 but with a release target of Java 8
to ensure backward compatibility.
