+++ .github/workflows/build.yml
@@ -15,6 +15,8 @@ jobs:
   Build-Setup:
     runs-on: ubuntu-latest
     env:
+      JAVA_DISTRIBUTION: temurin
+      JAVA_VERSION: 17
       ANYPOINT_CLIENT_ID: ${{ secrets.ANYPOINT_CLIENT_ID }}
       ANYPOINT_CLIENT_SECRET: ${{ secrets.ANYPOINT_CLIENT_SECRET }}
       ANYPOINT_CONNECTED_APP_ID: ${{ secrets.ANYPOINT_CONNECTED_APP_ID }}
@@ -48,6 +50,8 @@ jobs:
     secrets: inherit
     with:
       include-mule-ee-repo: true
+      java-distribution: ${{ env.JAVA_DISTRIBUTION }}
+      java-version: ${{ env.JAVA_VERSION }}
       maven-args: -Dtest=${{ needs.Build-Setup.outputs.test-suite }} -Danypoint.client.id=${{ needs.Build-Setup.outputs.ANYPOINT_CLIENT_ID }} -Danypoint.client.secret=${{needs.Build-Setup.outputs.ANYPOINT_CLIENT_SECRET}} -Danypoint.connected-app.id=${{needs.Build-Setup.outputs.ANYPOINT_CONNECTED_APP_ID}} -Danypoint.connected-app.secret=${{needs.Build-Setup.outputs.ANYPOINT_CONNECTED_APP_SECRET}} -Danypoint.username=${{needs.Build-Setup.outputs.ANYPOINT_USERNAME}} -Danypoint.password=${{needs.Build-Setup.outputs.ANYPOINT_PASSWORD}}
 
   Release-Maven:
@@ -57,8 +61,8 @@ jobs:
     with:
       app-version: ${{ needs.Build-Maven.outputs.app-version }}
       publish-maven-central: true
-    #   java-distribution: adopt-hotspot
-    #   java-version: 8
+      java-distribution: ${{ env.JAVA_DISTRIBUTION }}
+      java-version: ${{ env.JAVA_VERSION }}
     #   maven-args: -X
     #   main-branch: main
 
@@ -68,8 +72,8 @@ jobs:
     secrets: inherit
     with:
       app-version: ${{ needs.Build-Maven.outputs.app-version }}
-    #   java-distribution: adopt-hotspot
-    #   java-version: 8
+      java-distribution: ${{ env.JAVA_DISTRIBUTION }}
+      java-version: ${{ env.JAVA_VERSION }}
     #   maven-args: -X
     #   main-branch: main
     #   pr-reviewers: adesjardin, manikmagar, kkingavio
