+++ cli/pom.xml
@@ -10,7 +10,7 @@
     <parent>
         <groupId>com.avioconsulting.mule</groupId>
         <artifactId>mule-deploy-parent</artifactId>
-        <version>2.0.1-SNAPSHOT</version>
+        <version>2.1.0-SNAPSHOT</version>
     </parent>
 
     <build>
@@ -66,25 +66,6 @@
                 </configuration>
             </plugin>
             <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                    <compilerId>groovy-eclipse-compiler</compilerId>
-                </configuration>
-                <dependencies>
-                    <dependency>
-                        <groupId>org.codehaus.groovy</groupId>
-                        <artifactId>groovy-eclipse-compiler</artifactId>
-                        <version>3.7.0</version>
-                    </dependency>
-                    <dependency>
-                        <groupId>org.codehaus.groovy</groupId>
-                        <artifactId>groovy-eclipse-batch</artifactId>
-                        <version>${groovy.compiler.version}-03</version>
-                    </dependency>
-                </dependencies>
-            </plugin>
-            <plugin>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>${maven.surefire.plugin.version}</version>
                 <configuration>
@@ -120,11 +101,6 @@
             <version>4.7.4</version>
         </dependency>
         <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy</artifactId>
-            <version>${groovy.compiler.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.avioconsulting.mule</groupId>
             <artifactId>mule-deploy-library</artifactId>
             <version>${project.version}</version>
@@ -133,12 +109,6 @@
             <groupId>org.hamcrest</groupId>
             <artifactId>hamcrest-library</artifactId>
             <version>2.2</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy-test</artifactId>
-            <version>${groovy.compiler.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -159,15 +129,4 @@
         </repository>
     </repositories>
 
-    <pluginRepositories>
-        <pluginRepository>
-            <id>groovy</id>
-            <name>Groovy</name>
-            <layout>default</layout>
-            <url>https://groovy.jfrog.io/artifactory/libs-release-local</url>
-            <snapshots>
-
-            </snapshots>
-        </pluginRepository>
-    </pluginRepositories>
 </project>
