+++ library/pom.xml
@@ -10,7 +10,7 @@
     <parent>
         <groupId>com.avioconsulting.mule</groupId>
         <artifactId>mule-deploy-parent</artifactId>
-        <version>2.0.1-SNAPSHOT</version>
+        <version>2.1.0-SNAPSHOT</version>
     </parent>
 
     <properties>
@@ -33,25 +33,6 @@
                         </goals>
                     </execution>
                 </executions>
-            </plugin>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                    <compilerId>groovy-eclipse-compiler</compilerId>
-                </configuration>
-                <dependencies>
-                    <dependency>
-                        <groupId>org.codehaus.groovy</groupId>
-                        <artifactId>groovy-eclipse-compiler</artifactId>
-                        <version>3.7.0</version>
-                    </dependency>
-                    <dependency>
-                        <groupId>org.codehaus.groovy</groupId>
-                        <artifactId>groovy-eclipse-batch</artifactId>
-                        <version>${groovy.compiler.version}-03</version>
-                    </dependency>
-                </dependencies>
             </plugin>
             <plugin>
                 <groupId>com.bluetrainsoftware.maven</groupId>
@@ -184,6 +165,12 @@
                         <artifactId>velocity-engine-core</artifactId>
                         <version>2.1</version>
                     </dependency>
+<!--                Java 17 support https://github.com/groovy/gmaven/issues/28-->
+                    <dependency>
+                        <groupId>org.codehaus.groovy</groupId>
+                        <artifactId>groovy</artifactId>
+                        <version>3.0.13</version>
+                    </dependency>
                 </dependencies>
             </plugin>
             <!-- Right now pitest does not fully support Groovy but it can be slightly informative anyways -->
@@ -244,40 +231,23 @@
         </dependency>
         <dependency>
             <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy</artifactId>
-            <version>${groovy.compiler.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.groovy</groupId>
             <artifactId>groovy-json</artifactId>
-            <version>${groovy.compiler.version}</version>
+            <version>${groovy.version}</version>
         </dependency>
         <dependency>
             <groupId>org.codehaus.groovy</groupId>
             <artifactId>groovy-xml</artifactId>
-            <version>${groovy.compiler.version}</version>
+            <version>${groovy.version}</version>
         </dependency>
         <dependency>
             <groupId>org.codehaus.groovy</groupId>
             <artifactId>groovy-nio</artifactId>
-            <version>${groovy.compiler.version}</version>
+            <version>${groovy.version}</version>
         </dependency>
         <dependency>
             <groupId>org.mule.apikit</groupId>
             <artifactId>parser-service</artifactId>
             <version>2.1.2</version>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy-test</artifactId>
-            <version>${groovy.compiler.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy-ant</artifactId>
-            <version>${groovy.compiler.version}</version>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
@@ -372,16 +342,4 @@
             <layout>default</layout>
         </repository>
     </repositories>
-
-    <pluginRepositories>
-        <pluginRepository>
-            <id>groovy</id>
-            <name>Groovy</name>
-            <layout>default</layout>
-            <url>https://groovy.jfrog.io/artifactory/libs-release-local</url>
-            <snapshots>
-
-            </snapshots>
-        </pluginRepository>
-    </pluginRepositories>
 </project>
