+++ library/src/main/java/com/avioconsulting/mule/deployment/api/models/deployment/FileBasedAppDeploymentRequest.groovy
@@ -56,8 +56,7 @@ abstract class FileBasedAppDeploymentRequest extends AppDeploymentRequest {
     @Lazy
     protected PomInfo parsedPomProperties = {
         def zipOrJarPath = getFile().toPath()
-        FileSystems.newFileSystem(zipOrJarPath,
-                                  null).withCloseable { fs ->
+        FileSystems.newFileSystem(zipOrJarPath).withCloseable { fs ->
             def pomXmlPath = Files.walk(fs.getPath('/META-INF/maven')).find { p ->
                 p.endsWith('pom.xml')
             } as Path
@@ -92,8 +91,7 @@ abstract class FileBasedAppDeploymentRequest extends AppDeploymentRequest {
     @Override
     List<RamlFile> getRamlFilesFromApp(String rootRamlDirectory,
                                        boolean ignoreExchange) {
-        return FileSystems.newFileSystem(file.toPath(),
-                                         null).withCloseable { fs ->
+        return FileSystems.newFileSystem(file.toPath()).withCloseable { fs ->
             def apiPath = fs.getPath(rootRamlDirectory)
             if (!Files.exists(apiPath)) {
                 return []
