+++ library/src/test/java/com/avioconsulting/mule/deployment/internal/subdeployers/ApiManagerDeployerTest.groovy
@@ -24,7 +24,7 @@ import static org.junit.Assert.assertEquals
 import static org.junit.Assert.assertFalse
 import static org.junit.Assert.assertTrue
 
-@SuppressWarnings('GroovyAccessibility')
+@SuppressWarnings('groovy:access')
 class ApiManagerDeployerTest extends BaseTest {
     private ApiManagerDeployer deployer
