+++ library/src/test/java/com/avioconsulting/mule/deployment/internal/subdeployers/PolicyDeployerTest.groovy
@@ -18,7 +18,7 @@ import static org.hamcrest.MatcherAssert.assertThat
 import static org.hamcrest.Matchers.equalTo
 import static org.hamcrest.Matchers.is
 
-@SuppressWarnings('GroovyAccessibility')
+@SuppressWarnings('groovy:access')
 class PolicyDeployerTest extends BaseTest {
     private PolicyDeployer policyDeployer
