+++ library/src/test/resources/log4j2-test.xml
@@ -7,7 +7,7 @@
     </Appenders>
 
     <Loggers>
-        <Logger name="org.apache.http.wire" level="DEBUG"/>
+        <Logger name="org.apache.http.wire" level="INFO"/>
 
         <Root level="INFO">
             <AppenderRef ref="Console"/>
