+++ library/src/test/resources/mule4_project/mule-artifact.json
@@ -1,3 +1,3 @@
 {
-  "minMuleVersion": "4.1.4"
+  "minMuleVersion": "4.3.4"
 }
