+++ maven-plugin/pom.xml
@@ -10,7 +10,7 @@
     <parent>
         <groupId>com.avioconsulting.mule</groupId>
         <artifactId>mule-deploy-parent</artifactId>
-        <version>2.0.1-SNAPSHOT</version>
+        <version>2.1.0-SNAPSHOT</version>
     </parent>
 
 
@@ -52,25 +52,6 @@
                 </executions>
             </plugin>
             <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                    <compilerId>groovy-eclipse-compiler</compilerId>
-                </configuration>
-                <dependencies>
-                    <dependency>
-                        <groupId>org.codehaus.groovy</groupId>
-                        <artifactId>groovy-eclipse-compiler</artifactId>
-                        <version>3.7.0</version>
-                    </dependency>
-                    <dependency>
-                        <groupId>org.codehaus.groovy</groupId>
-                        <artifactId>groovy-eclipse-batch</artifactId>
-                        <version>${groovy.compiler.version}-03</version>
-                    </dependency>
-                </dependencies>
-            </plugin>
-            <plugin>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>${maven.surefire.plugin.version}</version>
                 <configuration>
@@ -106,11 +87,6 @@
             <version>3.3.9</version>
         </dependency>
         <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy</artifactId>
-            <version>${groovy.compiler.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.avioconsulting.mule</groupId>
             <artifactId>mule-deploy-library</artifactId>
             <version>${project.version}</version>
@@ -122,38 +98,11 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy-test</artifactId>
-            <version>${groovy.compiler.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>org.apache.maven.shared</groupId>
             <artifactId>maven-invoker</artifactId>
             <version>3.0.1</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
-
-    <repositories>
-        <repository>
-            <id>android</id>
-            <url>https://dl.bintray.com/apollographql/android/</url>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-        </repository>
-    </repositories>
-
-    <pluginRepositories>
-        <pluginRepository>
-            <id>groovy</id>
-            <name>Groovy</name>
-            <layout>default</layout>
-            <url>https://groovy.jfrog.io/artifactory/libs-release-local</url>
-            <snapshots>
-
-            </snapshots>
-        </pluginRepository>
-    </pluginRepositories>
+    
 </project>
