+++ maven-plugin/src/test/resources/mule4_project/pom.xml
@@ -1,65 +1,65 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
 
-	<groupId>f2ea2cb4-c600-4bb5-88e8-e952ff5591ee</groupId>
-	<artifactId>mule-deploy-lib-v4-test-app</artifactId>
+    <groupId>f2ea2cb4-c600-4bb5-88e8-e952ff5591ee</groupId>
+    <artifactId>mule-deploy-lib-v4-test-app</artifactId>
     <version>replaceme</version>
     <packaging>mule-application</packaging>
 
-	<name>mule-deploy-lib-v4-test-app</name>
+    <name>mule-deploy-lib-v4-test-app</name>
 
     <properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-		<app.runtime>replaceme</app.runtime>
-		<mule.maven.plugin.version>3.5.4</mule.maven.plugin.version>
-	</properties>
+        <app.runtime>replaceme</app.runtime>
+        <mule.maven.plugin.version>4.3.0</mule.maven.plugin.version>
+    </properties>
 
-	<build>
-		<plugins>
+    <build>
+        <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-deploy-plugin</artifactId>
-                <version>2.8.2</version>
+                <version>3.1.4</version>
                 <configuration>
                     <!-- parent project is only used during the build process -->
                     <skip>true</skip>
                 </configuration>
             </plugin>
-			<plugin>
-				<groupId>org.mule.tools.maven</groupId>
-				<artifactId>mule-maven-plugin</artifactId>
-				<version>${mule.maven.plugin.version}</version>
-				<extensions>true</extensions>
-				<configuration>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
+            <plugin>
+                <groupId>org.mule.tools.maven</groupId>
+                <artifactId>mule-maven-plugin</artifactId>
+                <version>${mule.maven.plugin.version}</version>
+                <extensions>true</extensions>
+                <configuration>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 
-	<dependencies>
+    <dependencies>
         <dependency>
             <groupId>org.mule.connectors</groupId>
             <artifactId>mule-http-connector</artifactId>
-            <version>1.3.2</version>
+            <version>1.10.3</version>
             <classifier>mule-plugin</classifier>
         </dependency>
         <dependency>
             <groupId>org.mule.connectors</groupId>
             <artifactId>mule-sockets-connector</artifactId>
-            <version>1.1.2</version>
+            <version>1.2.5</version>
             <classifier>mule-plugin</classifier>
         </dependency>
-	</dependencies>
+    </dependencies>
 
-	<repositories>
-          <repository>
-            <id>anypoint-exchange</id>
+    <repositories>
+        <repository>
+            <id>anypoint-exchange-v3</id>
             <name>Anypoint Exchange</name>
-            <url>https://maven.anypoint.mulesoft.com/api/v1/maven</url>
+            <url>https://maven.anypoint.mulesoft.com/api/v3/maven</url>
             <layout>default</layout>
         </repository>
         <repository>
